Bug Fixes and Changes
Replaced all instances of sprintf() with strlcpy(). 
Added an assertion to the strlcpy() to sanity/bounds check.
Added a check in the thread runner to avoid segfault and remediate issues when a drive is added/removed.
Re-initialize the LEDs when a disk is added/removed - should eliminate lighting issues.
Moved system led setting from main() to run_mediasmart(). This will set the system leds every time initialization occurs.